+++ src/multi_bspline_eval_sse_z_impl.h
@@ -85,7 +85,7 @@ eval_multi_UBspline_1d_z (multi_UBspline_1d_z *spline,
   a[2]  = (Ad[ 8]*tpx[0] + Ad[ 9]*tpx[1] + Ad[10]*tpx[2] + Ad[11]*tpx[3]);
   a[3]  = (Ad[12]*tpx[0] + Ad[13]*tpx[1] + Ad[14]*tpx[2] + Ad[15]*tpx[3]);
 
-  int xs = spline->x_stride;
+  intptr_t xs = spline->x_stride;
 
 
 
@@ -132,7 +132,7 @@ eval_multi_UBspline_1d_z_vg (multi_UBspline_1d_z *spline,
   da[2] = (dAd[ 8]*tpx[0] + dAd[ 9]*tpx[1] + dAd[10]*tpx[2] + dAd[11]*tpx[3]);
   da[3] = (dAd[12]*tpx[0] + dAd[13]*tpx[1] + dAd[14]*tpx[2] + dAd[15]*tpx[3]);
 
-  int xs = spline->x_stride;
+  intptr_t xs = spline->x_stride;
 
   for (int n=0; n<spline->num_splines; n++) {
     vals[n]  = 0.0;
@@ -242,8 +242,8 @@ eval_multi_UBspline_2d_z (multi_UBspline_2d_z *spline,
   tx = modf (ux, &ipartx);  int ix = (int) ipartx;
   ty = modf (uy, &iparty);  int iy = (int) iparty;
   
-  int xs = spline->x_stride;
-  int ys = spline->y_stride;
+  intptr_t xs = spline->x_stride;
+  intptr_t ys = spline->y_stride;
   int N  = spline->num_splines;
 
   // Now compute the vectors:
@@ -325,8 +325,8 @@ eval_multi_UBspline_2d_z_vg (multi_UBspline_2d_z *spline,
   tx = modf (ux, &ipartx);  int ix = (int) ipartx;
   ty = modf (uy, &iparty);  int iy = (int) iparty;
   
-  int xs = spline->x_stride;
-  int ys = spline->y_stride;
+  intptr_t xs = spline->x_stride;
+  intptr_t ys = spline->y_stride;
   int N  = spline->num_splines;
 
   // Now compute the vectors:
@@ -439,8 +439,8 @@ eval_multi_UBspline_2d_z_vgl (multi_UBspline_2d_z *spline,
   tx = modf (ux, &ipartx);  int ix = (int) ipartx;
   ty = modf (uy, &iparty);  int iy = (int) iparty;
   
-  int xs = spline->x_stride;
-  int ys = spline->y_stride;
+  intptr_t xs = spline->x_stride;
+  intptr_t ys = spline->y_stride;
   int N  = spline->num_splines;
 
   // Now compute the vectors:
@@ -570,8 +570,8 @@ eval_multi_UBspline_2d_z_vgh (multi_UBspline_2d_z *spline,
   tx = modf (ux, &ipartx);  int ix = (int) ipartx;
   ty = modf (uy, &iparty);  int iy = (int) iparty;
   
-  int xs = spline->x_stride;
-  int ys = spline->y_stride;
+  intptr_t xs = spline->x_stride;
+  intptr_t ys = spline->y_stride;
   int N  = spline->num_splines;
 
   // Now compute the vectors:
@@ -701,9 +701,9 @@ eval_multi_UBspline_3d_z (multi_UBspline_3d_z *spline,
   ty = modf (uy, &iparty);  int iy = (int) iparty;
   tz = modf (uz, &ipartz);  int iz = (int) ipartz;
   
-  int xs = spline->x_stride;
-  int ys = spline->y_stride;
-  int zs = spline->z_stride;
+  intptr_t xs = spline->x_stride;
+  intptr_t ys = spline->y_stride;
+  intptr_t zs = spline->z_stride;
   int N  = spline->num_splines;
 
   // Now compute the vectors:
@@ -850,9 +850,9 @@ eval_multi_UBspline_3d_z_vg (multi_UBspline_3d_z *spline,
   ty = modf (uy, &iparty);  int iy = (int) iparty;
   tz = modf (uz, &ipartz);  int iz = (int) ipartz;
   
-  int xs = spline->x_stride;
-  int ys = spline->y_stride;
-  int zs = spline->z_stride;
+  intptr_t xs = spline->x_stride;
+  intptr_t ys = spline->y_stride;
+  intptr_t zs = spline->z_stride;
   int N  = spline->num_splines;
 
   // Now compute the vectors:
@@ -991,9 +991,9 @@ eval_multi_UBspline_3d_z_vgl (multi_UBspline_3d_z *spline,
   ty = modf (uy, &iparty);  int iy = (int) iparty;
   tz = modf (uz, &ipartz);  int iz = (int) ipartz;
   
-  int xs = spline->x_stride;
-  int ys = spline->y_stride;
-  int zs = spline->z_stride;
+  intptr_t xs = spline->x_stride;
+  intptr_t ys = spline->y_stride;
+  intptr_t zs = spline->z_stride;
   int N  = spline->num_splines;
 
   // Now compute the vectors:
@@ -1156,9 +1156,9 @@ eval_multi_UBspline_3d_z_vgh (multi_UBspline_3d_z *spline,
   ty = modf (uy, &iparty);  int iy = (int) iparty;
   tz = modf (uz, &ipartz);  int iz = (int) ipartz;
   
-  int xs = spline->x_stride;
-  int ys = spline->y_stride;
-  int zs = spline->z_stride;
+  intptr_t xs = spline->x_stride;
+  intptr_t ys = spline->y_stride;
+  intptr_t zs = spline->z_stride;
   int N  = spline->num_splines;
 
   // Now compute the vectors:
